+++ utilities.py
@@ -221,8 +221,8 @@ def save_roi_signals(
         id_ = ids[i] if ids else i + 1
 
         roi_attrs = {
-            defs.DoricFile.Attribute.ROI.ID:           id_,
-            defs.DoricFile.Attribute.ROI.SHAPE:        0,
+            defs.DoricFile.Attribute.ROI.ID:           np.int32(id_),
+            defs.DoricFile.Attribute.ROI.SHAPE:        np.int32(0),
             defs.DoricFile.Attribute.ROI.COORDS:       footprint_to_coords(footprint),
             defs.DoricFile.Attribute.Dataset.USERNAME: usernames[i] if usernames else defs.DoricFile.Dataset.ROI.format(id_)
         }
